pyproject.toml
[project]
name = "claudshi"
version = "0.1.0"
description = "Political event prediction plugin for Claude using Kalshi prediction markets"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
